# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The interface documentation of the front and back end separate permission management system based on Gin + Vue + Element UI\nAdd qq group: 74520518 Enter the technical exchange group Please note, thank you!",
        "title": "go-admin API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/wenjianzhang/go-admin/blob/master/LICENSE.md"
        },
        "version": "0.0.1"
    },
    "paths": {
...
        "/login": {
            "post": {
                "description": "get token",
                "consumes": [
                    "application/json"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "description": "Add account",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"expire\": \"2019-08-07T12:45:48+08:00\", \"token\": \".eyJleHAiOjE1NjUxNTMxNDgsImlkIjoiYWRtaW4iLCJvcmlnX2lhdCI6MTU2NTE0OTU0OH0.-zvzHvbg0A\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "security": [
                    {
                        "": []
                    }
                ],
                "description": "获取token",
                "consumes": [
                    "application/json"
                ],
                "summary": "退出登录",
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"msg\": \"成功退出系统\" }",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Login": {
            "type": "object",
            "required": [
                "code",
                "password",
                "username",
                "uuid"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}